Move <:<, =:=, DummyImplicits out of Predef
There doesn't seem to be any good reason for them to be defined there.

<:< and =:= are kept in the same file since =:= extends from <:< and <:<
is sealed.

I didn't touch $conforms because I trust what the comment above it says,
but I'd like to know what the magic is exactly.
